Frequencies are now clipped to stay within safe ranges.
git-svn-id: http://ttblue.googlecode.com/svn/branches/active@266 bc2ad3ca-9a4d-0410-8d14-35da9c29b421

+++ source/units/TTBandpassButterworth.cpp
@@ -85,7 +85,7 @@ TTErr TTBandpassButterworth::clear()
 
 TTErr TTBandpassButterworth::setFrequency(TTValue& newValue)
 {
-	attrFrequency = clip(newValue, 10., sr*0.45);
+	attrFrequency = clip((double)newValue, 10., (sr*0.45));
 	
 	return calculateCoefficients();
 }
